+++ CHANGELOG.md
@@ -1,1 +1,14 @@
-# Changelog
+# v0.3.0 - Update Checker
+
+## Update Checker
+I have added a simple update checker so that you can always be notified when there is
+a new release! It will notify the staff with the permission detoxify.admin and notify
+console when the server is started.
+
+This feature can be disabled within the config by disabling `check-updates`
+
+---
+
+## Other Things
+I aso removed some annoying debug messages that were left over from me testing the
+config system. As well as some other non-necessary messages